refactor: move pages to presentation layer
