--- app/(protected)/extract/index.tsx
@@ -1,5 +1,0 @@
-import DinamicTransactionsList from "@/components/transactions/DinamicTransactionsList";
-
-export default function Extract() {
-  return <DinamicTransactionsList />;
-}
