+++ app/screens/(auth)/register/style.ts
