+++ app/screens/(auth)/splash-screen/style.ts
