+++ app/screens/(protected)/extract/index.tsx
@@ -1,0 +1,5 @@
+import DinamicTransactionsList from "@/app/components/transactions/DinamicTransactionsList";
+
+export default function Extract() {
+  return <DinamicTransactionsList />;
+}
